logo create tabelle

+++ database/migrations/2024_11_29_123028_create_logos.php
@@ -11,14 +11,16 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('logos', function (Blueprint $table) {
-            $table->id(); // Primärschlüssel
-            $table->string('name')->comment('Name des Logos');
-            $table->string('filename')->comment('Name der Datei');
-            $table->string('localfilename')->comment('Name der Datei');
-            $table->foreignId('client_id')->constrained('clients')->onDelete('cascade');
-            $table->timestamps(); // Erstellt automatisch created_at und updated_at
-        });
+        if (!Schema::hasTable('logos')) {
+            Schema::create('logos', function (Blueprint $table) {
+                $table->id(); // Primärschlüssel
+                $table->string('name')->comment('Name des Logos');
+                $table->string('filename')->comment('Name der Datei');
+                $table->string('localfilename')->comment('Name der Datei');
+                $table->foreignId('client_id')->constrained('clients')->onDelete('cascade');
+                $table->timestamps(); // Erstellt automatisch created_at und updated_at
+            });
+        }
     }
 
     /**
